requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
geoalchemy2
pydantic
geoalchemy2[shapely]
xmltodict
geopy
requests
python-dotenv
google-genai
newsapi-python
gdacs-api
tqdm